CM-852: clean up the project
