+++ dotnet-efcore/SingleAndSplitQueriesInEFCore/SingleAndSplitQueriesInEFCore/Program.cs
@@ -5,19 +5,4 @@ using SingleAndSplitQueriesInEFCore;
 await using var db = new CompaniesContext();
 await db.Database.MigrateAsync();
 
-
-//await db.Companies
-//    .Include(company => company.Products)
-//    .Include(company => company.Departments)
-//    .ThenInclude(department => department.Employees)
-//    .AsSingleQuery()
-//    .ToListAsync();
-
-//await db.Companies
-//    .Include(company => company.Products)
-//    .Include(company => company.Departments)
-//    .ThenInclude(department => department.Employees)
-//    .AsSplitQuery()
-//    .ToListAsync();
-
 BenchmarkRunner.Run<SingleAndSplitBenchmark>();
