+++ dotnet-efcore/SingleAndSplitQueriesInEFCore/SingleAndSplitQueriesInEFCore/SingleAndSplitQueriesInEFCore.csproj
@@ -10,7 +10,6 @@
   <ItemGroup>
     <PackageReference Include="BenchmarkDotNet" Version="0.13.12" />
     <PackageReference Include="Bogus" Version="35.4.0" />
-    <PackageReference Include="EFCore.BulkExtensions.SqlServer" Version="8.0.1" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="8.0.1">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
